tests: move method userShouldGetSSEEvent from api to ui
